assert: remove build tag for ErrorIs
Go 1.12 (the last version before errors.Is was added) has not been
supported in a while.

Move ErrorIs assertion and comparison into the main files.
