+++ go.mod
@@ -1,5 +1,7 @@
 module gotest.tools/v3
 
+go 1.13
+
 require (
 	github.com/google/go-cmp v0.5.5
 	github.com/pkg/errors v0.9.1
@@ -7,5 +9,3 @@ require (
 	golang.org/x/sys v0.0.0-20210119212857-b64e53b001e4
 	golang.org/x/tools v0.1.0
 )
-
-go 1.11
